requirements.txt
streamlit==1.29.0
supabase==2.0.4
pandas==2.1.4
bcrypt==4.1.2
pytz==2023.3
PyYAML==6.0.1
